requirements.txt
discord.py
python-dotenv
pynacl
yt-dlp
ffmpeg